+++ src/features/chat/hooks/chatList/useChatList.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
-import { CHAT_KEYS } from '../constants/queryKeys';
-import { getChatList } from '../api/chats/getChatList';
-import { ChatListItem } from '../types';
+import { ChatListItem } from '../../types';
+import { CHAT_KEYS } from '../../constants/queryKeys';
+import { getChatList } from '../../api/chats/getChatList';
 
 export const useChatList = (limit = 10) => {
   return useQuery<ChatListItem[]>({
